Finalizando desafio CRUD Clientes

+++ src/main/java/com/devsuperior/DesafioCrudCliente/services/ClientService.java
@@ -1,7 +1,7 @@
 package com.devsuperior.DesafioCrudCliente.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
-
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -9,6 +9,12 @@ import org.springframework.stereotype.Service;
 import com.devsuperior.DesafioCrudCliente.dto.ClientDTO;
 import com.devsuperior.DesafioCrudCliente.entities.Client;
 import com.devsuperior.DesafioCrudCliente.repositories.ClientRepository;
+import com.devsuperior.DesafioCrudCliente.services.exceptions.DatabaseException;
+import com.devsuperior.DesafioCrudCliente.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
+
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -19,8 +25,10 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        ClientDTO client = repository.findById(id).map(c -> new ClientDTO(c)).get();
-        return client;
+        Client client = repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
+
+        return new ClientDTO(client);
     }
 
     @Transactional(readOnly = true)
@@ -38,23 +46,35 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
         entity.setName(dto.getName());
-        entity.setCpf(dto.getCpf());   
+        entity.setCpf(dto.getCpf());
         entity.setIncome(dto.getIncome());
-        entity.setBirthDate(dto.getBirthDate()); 
-        entity.setChildren(dto.getChildren());  
+        entity.setBirthDate(dto.getBirthDate());
+        entity.setChildren(dto.getChildren());
     }
 
 }
